+++ components/dashboard/Assets.tsx
@@ -1,19 +1,32 @@
 import { useNavigate } from "react-router-dom";
 import kasIcon from "@/assets/images/kas-icon.svg";
-import { formatToken, formatTokenPrice, formatUSD } from "@/lib/utils.ts";
+import {
+  formatCurrency,
+  formatToken,
+  formatTokenPrice,
+  symbolForCurrencyCode,
+} from "@/lib/utils.ts";
 import TokenListItem from "@/components/dashboard/TokenListItem.tsx";
 import { applyDecimal } from "@/lib/krc20.ts";
+import useCurrencyValue from "@/hooks/useCurrencyValue.ts";
 
 export default function Assets() {
   const navigate = useNavigate();
-  const kapsaPrice = useKaspaPrice();
+  const kaspaPrice = useKaspaPrice();
 
   const [settings] = useSettings();
   const { account } = useWalletManager();
 
   const address = account?.address;
   const balance = account?.balance;
+  const fiatBalance = parseFloat(balance ?? "0") * kaspaPrice.kaspaPrice;
   const showBalance = !settings?.hideBalances;
+  const fiatKaspaPrice = kaspaPrice.kaspaPrice;
+
+  const { amount: totalBalanceCurrency, code: currencyCode } =
+    useCurrencyValue(fiatBalance);
+  const { amount: kaspaPriceCurrency, code: kaspaPriceCurrencyCode } =
+    useCurrencyValue(fiatKaspaPrice);
 
   const { data: tokenListResponse } = useTokenListByAddress(address, 5000);
   const tokenListItems = tokenListResponse?.result
@@ -51,13 +64,14 @@ export default function Assets() {
             </span>
           </div>
           <div className="flex items-center justify-between text-sm text-daintree-400">
-            <span>{formatTokenPrice(kapsaPrice.kaspaPrice)}</span>
+            <span>
+              {formatTokenPrice(kaspaPriceCurrency, kaspaPriceCurrencyCode)}
+            </span>
             <span>
               ≈{" "}
               {showBalance
-                ? formatUSD(parseFloat(balance ?? "0") * kapsaPrice.kaspaPrice)
-                : "$*****"}{" "}
-              USD
+                ? formatCurrency(totalBalanceCurrency, currencyCode)
+                : `${symbolForCurrencyCode(currencyCode)}*****`}
             </span>
           </div>
         </div>